requirements.txt
torch==2.2.0
torchvision==0.17.0
torchmetrics
numpy==1.26.4
tqdm
matplotlib
pandas
wandb
hydra-core
linear_attention_transformer
scikit-learn

black
isort
pre-commit
